Fix incorect mask blending (now based on black masks). Some minor bug fixes (prevent negative depth).
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@100 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/AlgorithmSelectionDialog.cpp
@@ -188,7 +188,7 @@ int AlgorithmSelectionDialog::getSelectedHeight(){
 
 unsigned long  AlgorithmSelectionDialog::getUpdatePeriod(){
 
-	return (unsigned long) ( 1000000.f / float(frequencySlider->value()));
+	return (unsigned long) ( 1000000.0 / double(frequencySlider->value()));
 
 }
